refactor get all public post
